missed another. this is what happens when you casually use msvc and attempt partial commits...
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4488 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/sbar.c
@@ -227,7 +227,7 @@ static mpic_t *Sbar_Q2CachePic(char *name)
 #define	ICON_HEIGHT	24
 #define	CHAR_WIDTH	16
 #define	ICON_SPACE	8
-static void SCR_DrawField (float x, float y, int color, float width, float value)
+static void SCR_DrawField (float x, float y, int color, float width, int value)
 {
 	char	num[16], *ptr;
 	int		l;
